Improvements in content disposition and content length handling

+++ requests_downloader/downloader.py
@@ -11,7 +11,7 @@ import re
 import hashlib
 import logging
 import mimetypes
-from urllib.parse import urlparse, urlunparse
+from urllib.parse import urlparse, urlunparse, unquote
 
 import requests
 from tqdm import tqdm
@@ -43,6 +43,7 @@ def handle_url(url):
     drive_pattern_1 = rf'{drive}/file/d/([^\/]*)/.*'
     drive_match_1 = re.match(drive_pattern_1, url)
     if drive_match_1:
+        log.debug("Google Drive pattern-1 matched.")
         file_id = drive_match_1.group(1)
         dl_url = f'{drive}/u/0/uc?id={file_id}&export=download'
         return [('drive', dl_url)], default_idx
@@ -50,6 +51,7 @@ def handle_url(url):
     drive_pattern_2 = rf'{drive}/open\?id=([^\/&]*).*'
     drive_match_2 = re.match(drive_pattern_2, url)
     if drive_match_2:
+        log.debug("Google Drive pattern-2 matched.")
         file_id = drive_match_2.group(1)
         dl_url = f'{drive}/u/0/uc?id={file_id}&export=download'
         return [('drive', dl_url)], default_idx
@@ -63,8 +65,10 @@ def handle_url(url):
     doc_pattern = rf'{docs}/(spreadsheets|document|presentation)/d/([^\/]*)/.*'
     doc_match = re.match(doc_pattern, url)
     if doc_match:
+        log.debug("Google Docs pattern matched.")
         doc_type = doc_match.group(1)
         doc_id = doc_match.group(2)
+        log.debug("Type: {doc_type}, ID: {doc_id}")
         dl_types = preference[doc_type]
         dl_urls = [
             (dl_type, (f'{docs}/{doc_type}/d/{doc_id}/export?'
@@ -77,6 +81,7 @@ def handle_url(url):
     archive_pattern = rf'{archive}/(details|download)/([^\/]*).*'
     archive_match = re.match(archive_pattern, url)
     if archive_match:
+        log.debug("Archive.org pattern matched.")
         content_name = archive_match.group(2)
         archive_url = f'{archive}/download/{content_name}'
         dl_url = f'{archive}/compress/{content_name}'
@@ -122,6 +127,7 @@ def handle_url(url):
         parse_result = parse_result._replace(query=query_string)
         return [('dropbox', urlunparse(parse_result))], default_idx
 
+    log.debug("No specific pattern matched.")
     return [('direct', url)], default_idx
 
 
@@ -197,6 +203,7 @@ def download(url, download_dir='', download_file=None, download_path=None,
     log.debug(f"URL: {url}")
 
     request_maker = requests if session is None else session
+    headers['Range'] = 'bytes=0-'
 
     r = request_maker.head(url, headers=headers, timeout=timeout)
 
@@ -211,8 +218,23 @@ def download(url, download_dir='', download_file=None, download_path=None,
     content_length = int(r.headers.get('content-length', 0))
     log.debug(f"Content-Length: {content_length}")
 
+    content_range = r.headers.get('content-range', '')
+    _content_range_part = content_range.split('/')[-1].strip()
+    log.debug(f"Content-Range: {content_range}")
+
+    if content_length == 0 and _content_range_part:
+        content_length = int(_content_range_part)
+        log.debug(f"Content-Length (from Range): {content_length}")
+
     content_type = r.headers.get('content-type')
+    html_content = (content_type == 'text/html; charset=utf-8')
     log.debug(f"Content-Type: {content_type}")
+    log.debug(f"HTML Content: {html_content}")
+
+    if html_content:
+        log.error("HTML content detected.")
+        log.error(f"Download from {url} aborted.")
+        return False
 
     extension_guess = mimetypes.guess_extension(content_type)
     log.debug(f"Extension Guess: {extension_guess}")
@@ -220,17 +242,49 @@ def download(url, download_dir='', download_file=None, download_path=None,
     visible_name = r.url.split('/')[-1]
     if extension_guess and not visible_name.endswith(extension_guess):
         visible_name += f'.{extension_guess}'
+    log.debug(f"Visible Name: {visible_name}")
 
     provided_name = None
     cd = r.headers.get('content-disposition', None)
     log.debug(f"Content-Disposition: {cd}")
     if cd is not None:
-        provided_names = re.findall('filename="(.+)"', cd)
+        cd_fields = {}
+        for part in cd.split(';'):
+            _kv = part.split('=')
+            if '=' in part:
+                _k = _kv[0].strip(' \t\n"\'')
+                _v = _kv[1].strip(' \t\n"\'')
+                cd_fields[_k] = _v
+
+        log.debug(cd_fields)
+        provided_names = [
+            v for k, v in cd_fields.items() if k == 'filename'
+        ]
+        provided_encoded_names = [
+            v for k, v in cd_fields.items() if k == 'filename*'
+        ]
+
+        # provided_names = re.findall('filename="(.+)"', cd)
+        log.debug(f"Filenames: {provided_names}")
+        log.debug(f"Filenames*: {provided_encoded_names}")
+
         if provided_names:
             provided_name = provided_names[0]
 
+        if provided_encoded_names:
+            encoding, name = provided_encoded_names[0].split("''")
+            print(encoding, name)
+            provided_name = unquote(name, encoding=encoding)
+
+        log.debug(f"Final Provided Name: {provided_name}")
+
     if not download_file:
         download_file = provided_name if provided_name else visible_name
+
+    if not download_file:
+        log.error("Download location could not be inferred.")
+        log.error(f"Download from {url} aborted.")
+        return False
 
     if not download_path:
         download_path = os.path.join(download_dir, download_file)
@@ -244,6 +298,7 @@ def download(url, download_dir='', download_file=None, download_path=None,
 
     with open(download_path, 'ab') as f:
         position = f.tell()
+        log.debug(f"Current Position: {position}")
         if position and position == content_length:
             log.info(f"File '{download_file}' is already downloaded!")
             return download_path
@@ -273,12 +328,19 @@ def download(url, download_dir='', download_file=None, download_path=None,
                 wrote += f.write(data)
                 t.update(len(data))
 
+    log.debug(f"Wrote: {wrote}")
+
     if content_length == 0:
-        os.unlink(download_path)
-        log.warning(
-            f"Downloaded file '{download_file}' was empty and was removed."
-        )
-        success = False
+        filesize = os.stat(download_path).st_size
+        log.debug(f"Filesize: {filesize}")
+        if not filesize:
+            os.unlink(download_path)
+            log.warning(
+                f"Downloaded file '{download_file}' was empty and was removed."
+            )
+            success = False
+        else:
+            log.warning(f"Integrity of '{download_file}' could not verified.")
     elif (position + wrote) != content_length:
         success = False
         log.warning(f"Inconsistency in download from '{url}'.")
